add --auto-height option
When enabled, shrink the window height to match the number of input
items when reading from stdin rather than always clearing a sizeable
chunk of the screen.

The implementation is a bit hacky and I think this is technically
a semver break due to adding a pub field to the all-pub SkimOptions
struct, but I don't care.

This works by reading and buffering lines from stdin until we would have
filled the specified height or until EOF. If EOF comes before "height"
number of lines, we reduce the height option before calling into the
Skim library.