fix and update clap (part 2): update to clap v4.1
Now that clap 3.2 doesn't immediately assert in debug builds, update to
clap v4, which has better documentation for the builder-pattern, and
officially supports args_override_self.

The big change here is that Arg::takes_value() is gone, and that's
assumed to be the default (via ArgAction::Set).

Also now we need Command::disable_help_flag() to do the custom help
text, but I'll be ditching that soon too.