Revert "transparency on start"
What a terrible "feature", that seemingly just totally broke terminal
handling. Note that this also rolls tuikit back to v0.4.6, since this
change is tightly linked with that tuikit update.

It's a fractal of failure here, util::clear_canvas does weird painting
when ANSI colors are involved, and due to what I think is a bug in
tuikit v0.5.0, no combination of disable_alternate_screen and
clear_on_start work properly together in all cases: it always breaks
either when the window is "too big" with extra space, or when the window
is "too small" and skim takes up the whole screen.

I don't need any of this mess, and I don't feel like forking and
maintaining tuikit too, so just revert all of it.

This reverts commit 04d4544c2190879ca2ba54b1e8e53c4a893518f0.