WIP: use clap to generate help text
Ditch the custom separate help text and use clap instead. This is
actually kinda fine in clap v4 beacuse of next_help_heading (back in the
clap 2 days, I see why the skim authors opted for a custom help
message).

This is just a WIP, 10 arguments down, 58 more to go.