replace custom SpinLock with std::sync::Mutex
Seeing a hand-rolled spinlock implementation inside application code
does not spark joy. It's tough to believe it's meaningfully better than
an off-the-shelf Mutex type, such as std::sync::Mutex.

I don't have a way to benchmark performance differences, but if I don't
notice subjectively then it's probably fine. The Mutex in std has
supposedly gotten a lot faster lately when it switched to a Futex on
Linux rather than a wrapper around pthread_mutex_t.

Leave the old code in place, just hide it behind a "spinlock" cargo
feature, disabled by default. Rather than refactoring users, just
newtype and alias Mutex and MutexGuard so we can present the same API to
the rest of the crate.