replace lazy_static with once_cell
Most of the uses of lazy_static were for regexes so I made a lazy_regex!
macro at the crate root to help with that, which makes a Lazy<Regex>.

A few other uses were things that didn't need lazy_static or once_cell
at all and could just be normal constants or statics.

Note that lazy_static is still in the crate graph from tuikit (which I'm
not sure if I want to fork yet...)